Simplified ID derivation.
